requirements.txt
mysql-connector-python==8.0.33
tqdm==4.66.1
rich==13.7.0
psutil==5.9.8
tabulate==0.9.0 